Refactor some CSV related files
Rename the 'csv_handler'-folder to 'csv' and move the 'TimerHandler'-class into its own folder as well as moving the 'test_timer_handler'-file into its own folder
